分配端口区间

+++ aop-web-server/src/main/java/net/arksea/ansible/deploy/api/manage/dao/PortTypeDao.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Create by xiaohaixing on 2020/9/17
  */
-public interface PortTypeDao extends CrudRepository<PortType, Long> {
+public interface PortTypeDao extends CrudRepository<PortType, Integer> {
 }
